move GroupActivityItem into its own file
